+++ codec-http/src/main/java/io/netty/handler/codec/http/websocketx/WebSocketServerHandshaker08.java
@@ -62,7 +62,7 @@ public class WebSocketServerHandshaker08 extends WebSocketServerHandshaker {
      *            Maximum allowable frame payload length. Setting this value to your application's requirement may
      *            reduce denial of service attacks using long data frames.
      */
-    public WebSocketServerHandshaker08(String webSocketURL, String subprotocols, boolean allowExtensions, long maxFramePayloadLength) {
+    public WebSocketServerHandshaker08(String webSocketURL, String subprotocols, boolean allowExtensions, int maxFramePayloadLength) {
         super(WebSocketVersion.V08, webSocketURL, subprotocols, maxFramePayloadLength);
         this.allowExtensions = allowExtensions;
     }
